Hoist PDFSectorMap out of Map.py
